requirements.txt
python-dotenv
openai
ipykernel
requests